Optimización de codigo y tipado

+++ src/controllers/customer.controller.ts
@@ -13,19 +13,19 @@ export const signUp = async (req: Request, res: Response) => {
     try {
         const { firstname, lastname, email, phone, dni, age, password } = req.body;
         await signUpSchema.validateAsync(req.body);
-        const customer = new Customer();
+        const customer: Customer = new Customer();
         customer.firstname = firstname;
         customer.lastname = lastname;
         customer.password = bcrypt.hashSync(password, salt);
         customer.email = email;
         customer.phone = phone;
-        customer.dni = dni,
+        customer.dni = dni;
         customer.age = age;
         await customerRepository.save(customer);
-        return res.status(201).send({message: `¡Cuenta creada! Bienvenido ${customer.firstname} ${customer.lastname}`});
+        return res.status(201).send({ message: `¡Cuenta creada! Bienvenido ${customer.firstname} ${customer.lastname}` });
     } catch (error) {
         if (error instanceof Error) {
-            return res.status(500).send({message: error.message});
+            return res.status(500).send({ message: error.message });
         }
     }
 }
@@ -33,14 +33,14 @@ export const signUp = async (req: Request, res: Response) => {
 // GET
 export const getAllCustomers = async (req: Request, res: Response) => {
     try {
-        const customers = await customerRepository.find();
+        const customers: Customer[] = await customerRepository.find();
         if (customers.length === 0) {
-            return res.status(404).send({message: "No hay ningún cliente registrado aún"});
+            return res.status(404).send({ message: "No hay ningún cliente registrado aún" });
         }
-        return res.status(200).send({message: `Customers: `, customers});
+        return res.status(200).send({ message: `Customers: `, customers });
     } catch (error) {
         if (error instanceof Error) {
-            return res.status(500).send({message: error.message});
+            return res.status(500).send({ message: error.message });
         }
     }
 }
@@ -48,14 +48,14 @@ export const getAllCustomers = async (req: Request, res: Response) => {
 // GET
 export const getCustomerById = async (req: Request, res: Response) => {
     try {
-        const customer = await customerRepository.findOneBy({id: parseInt(req.params.id)});
-        if (!customer) {
-            return res.status(404).send({message: "No se encontró ningún cliente con ese ID"});
+        const customer: Customer | null = await customerRepository.findOneBy({ id: parseInt(req.params.id) });
+        if (!customer || customer === null) {
+            return res.status(404).send({ message: "No se encontró ningún cliente con ese ID" });
         }
-        return res.status(200).send({message: `Cliente #${req.params.id}`, customer});
+        return res.status(200).send({ message: `Cliente #${req.params.id}`, customer });
     } catch (error) {
         if (error instanceof Error) {
-            return res.status(500).send({message: error.message});
+            return res.status(500).send({ message: error.message });
         }
     }
 }
@@ -63,15 +63,15 @@ export const getCustomerById = async (req: Request, res: Response) => {
 // DELETE 
 export const deleteUserById = async (req: Request, res: Response) => {
     try {
-        const customer = await customerRepository.findOneBy({id: parseInt(req.params.id)});
-        if (!customer) {
-            return res.status(404).send({message: "No se encontró ningún cliente con ese ID"});
+        const customer: Customer | null = await customerRepository.findOneBy({ id: parseInt(req.params.id) });
+        if (!customer || customer === null) {
+            return res.status(404).send({ message: "No se encontró ningún cliente con ese ID" });
         }
         await customerRepository.delete(customer);
-        return res.status(200).send({message: `Cliente #${req.params.id} eliminado`});
+        return res.status(200).send({ message: `Cliente #${req.params.id} eliminado` });
     } catch (error) {
         if (error instanceof Error) {
-            return res.status(500).send({message: error.message});
+            return res.status(500).send({ message: error.message });
         }
     }
 }
@@ -81,19 +81,19 @@ export const logIn = async (req: Request, res: Response) => {
     try {
         const { email, password } = req.body;
         await logInSchema.validateAsync(req.body);
-        const customer = await customerRepository.findOne({where: {email: (email)}});
-        if (!customer) {
-            return res.status(404).send({message: `No se encontró ninguna cuenta con el email '${email}'. Intente nuevamente`})
+        const customer: Customer | null = await customerRepository.findOne({ where: { email: (email) } });
+        if (!customer || customer === null) {
+            return res.status(404).send({ message: `No se encontró ninguna cuenta con el email '${email}'. Intente nuevamente` })
         }
-        const validatePassw = await bcrypt.compare(password, customer.password);
+        const validatePassw: boolean = await bcrypt.compare(password, customer.password);
         if (!validatePassw) {
-            return res.status(404).send({message: "Contraseña incorrecta. Intente nuevamente"});
+            return res.status(404).send({ message: "Contraseña incorrecta. Intente nuevamente" });
         }
-        const token = await signToken(customer);
-        return res.status(200).send({message: "Ingreso exitoso. ¡Bienvenido!", token})
+        const token: string = await signToken(customer);
+        return res.status(200).send({ message: "Ingreso exitoso. ¡Bienvenido!", token })
     } catch (error) {
         if (error instanceof Error) {
-            return res.status(500).send({message: error.message});
+            return res.status(500).send({ message: error.message });
         }
     }
 }
